+++ NEWS.md
@@ -1,3 +1,7 @@
+# swfscDAS 0.2 (in development)
+
+* Renamed 'siteinfo' in `das_effort` and `das_effort_sight` output to `sightinfo`
+
 # swfscDAS 0.1
 
 * Initial version, presented to NMFS users